style(contentful): move & rename ContentsDelivery
