+++ packages/botonic-plugin-contentful/src/contentful/delivery-api.ts
@@ -1,22 +1,17 @@
 import * as contentful from 'contentful'
-import { CreateClientParams, Entry, EntryCollection } from 'contentful'
+import { CreateClientParams, Entry } from 'contentful'
 import {
   CmsException,
   CommonFields,
-  Content,
   ContentType,
   Context,
   FollowUp,
   SearchableBy,
-  TopContent,
-  TopContentType,
 } from '../cms'
-import { QueueDelivery } from './contents/queue'
 import {
   SearchableByKeywordsDelivery,
   SearchableByKeywordsFields,
 } from './search/searchable-by'
-import { ScheduleDelivery } from './contents/schedule'
 import { DateRangeDelivery, DateRangeFields } from './contents/date-range'
 import { ReducedClientApi } from './delivery/client-api'
 import { ContentfulOptions } from '../plugin'
@@ -82,61 +77,6 @@ export class AdaptorDeliveryApi implements DeliveryApi {
       query['locale'] = context.locale
     }
     return query
-  }
-}
-
-export class ContentsApi {
-  constructor(readonly api: DeliveryApi) {}
-
-  async contents(
-    contentType: ContentType,
-    context: Context,
-    factory: (entry: contentful.Entry<any>, ctxt: Context) => Promise<Content>
-  ): Promise<Content[]> {
-    const entryCollection: EntryCollection<CommonEntryFields> = await this.api.getEntries(
-      context,
-      {
-        // eslint-disable-next-line @typescript-eslint/naming-convention
-        content_type: contentType,
-        include: this.maxReferencesInclude(),
-      }
-    )
-    return Promise.all(
-      entryCollection.items.map(entry => factory(entry, context))
-    )
-  }
-
-  async topContents(
-    model: TopContentType,
-    context: Context,
-    factory: (
-      entry: contentful.Entry<any>,
-      ctxt: Context
-    ) => Promise<TopContent>,
-    filter?: (cf: CommonFields) => boolean
-  ): Promise<TopContent[]> {
-    const entryCollection: EntryCollection<CommonEntryFields> = await this.api.getEntries(
-      context,
-      {
-        // eslint-disable-next-line @typescript-eslint/naming-convention
-        content_type: model,
-        include: this.maxReferencesInclude(),
-      }
-    )
-    let entries = entryCollection.items
-    if (filter) {
-      entries = entries.filter(entry =>
-        filter(ContentfulEntryUtils.commonFieldsFromEntry(entry))
-      )
-    }
-    return Promise.all(entries.map(entry => factory(entry, context)))
-  }
-
-  private maxReferencesInclude() {
-    return Math.max(
-      QueueDelivery.REFERENCES_INCLUDE,
-      ScheduleDelivery.REFERENCES_INCLUDE
-    )
   }
 }
